Added Redis Caching and enhancements

+++ src/main/java/com/navya/hotelbookingservice/Booking.java
@@ -35,7 +35,7 @@ public class Booking {
     @JsonFormat(pattern = "yyyy-MM-dd")
     Date endDate;
 
-    Double totalPrice;
+    Integer totalPrice;
